requirements.txt
beautifulsoup4==4.12.2
dvc==3.30.3
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
